Fix a lot of warning

+++ src/filters/transform/vsfilter/DirectVobSubFilter.cpp
@@ -144,7 +144,7 @@ CDirectVobSubFilter::~CDirectVobSubFilter()
 	if(m_hbm) {DeleteObject(m_hbm); m_hbm = 0;}
 	if(m_hdc) {DeleteObject(m_hdc); m_hdc = 0;}
 
-	for(int i = 0; i < m_pTextInput.GetCount(); i++)
+	for(size_t i = 0; i < m_pTextInput.GetCount(); i++)
 		delete m_pTextInput[i];
 
 	m_frd.EndThreadEvent.Set();
@@ -1729,7 +1729,7 @@ bool CDirectVobSubFilter::Open()
 	CAtlArray<SubFile> ret;
 	GetSubFileNames(m_FileName, paths, ret);
 
-	for(int i = 0; i < ret.GetCount(); i++)
+	for(size_t i = 0; i < ret.GetCount(); i++)
 	{
 		if(m_frd.files.Find(ret[i].fn))
 			continue;
@@ -1777,7 +1777,7 @@ bool CDirectVobSubFilter::Open()
 		}
 	}
 
-	for(int i = 0; i < m_pTextInput.GetCount(); i++)
+	for(size_t i = 0; i < m_pTextInput.GetCount(); i++)
 	{
 		if(m_pTextInput[i]->IsConnected())
         {
@@ -1961,7 +1961,7 @@ void CDirectVobSubFilter::AddSubStream(ISubStream* pSubStream)
     }
 
 	int len = m_pTextInput.GetCount();
-	for(int i = 0; i < m_pTextInput.GetCount(); i++)
+	for(size_t i = 0; i < m_pTextInput.GetCount(); i++)
 		if(m_pTextInput[i]->IsConnected()) len--;
 
 	if(len == 0)
@@ -2015,7 +2015,7 @@ void CDirectVobSubFilter::SetupFRD(CStringArray& paths, CAtlArray<HANDLE>& handl
 {
     CAutoLock cAutolock(&m_csSubLock);
 
-	for(int i = 2; i < handles.GetCount(); i++)
+	for(size_t i = 2; i < handles.GetCount(); i++)
 	{
 		FindCloseChangeNotification(handles[i]);
 	}
@@ -2142,7 +2142,7 @@ DWORD CDirectVobSubFilter::ThreadProc()
 		}
 	}
 
-	for(int i = 2; i < handles.GetCount(); i++)
+	for(size_t i = 2; i < handles.GetCount(); i++)
 	{
 		FindCloseChangeNotification(handles[i]);
 	}
